Create export function from cli
